+++ actionpack/test/controller/parameters/parameters_permit_test.rb
@@ -520,9 +520,13 @@ class ParametersPermitTest < ActiveSupport::TestCase
     assert_equal false, params.permitted?
   end
 
-  test "only String and Symbol keys are allowed" do
+  test "only String, Symbol and Integer keys are allowed" do
     assert_raises(ActionController::InvalidParameterKey) do
       ActionController::Parameters.new({ foo: 1 } => :bar)
+    end
+
+    assert_nothing_raised do
+      ActionController::Parameters.new(4 => :bar)
     end
   end
 end